CLN: Split common file to interface + common
